# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Process = typeof Process[keyof typeof Process];


export const Process = {
  reliability: 'reliability',
  laser: 'laser',
  lkc: 'lkc',
  fuc: 'fuc',
  ct: 'ct',
  tag: 'tag',
} as const;

export interface ProcessState {
  done: boolean;
  /** @nullable */
  completedBy?: string | null;
  /** @nullable */
  completedAt?: string | null;
}

export type MachineStatus = typeof MachineStatus[keyof typeof MachineStatus];


export const MachineStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type MachineProcesses = {
  reliability: ProcessState;
  laser: ProcessState;
  lkc: ProcessState;
  fuc: ProcessState;
  ct: ProcessState;
  tag: ProcessState;
};

export interface Machine {
  id: number;
  bedNumber: string;
  name: string;
  status: MachineStatus;
  processes: MachineProcesses;
  createdAt: string;
  updatedAt: string;
}

export interface MachineInput {
  /** @minLength 1 */
  bedNumber: string;
  /** @minLength 1 */
  name: string;
}

export interface MachineUpdate {
  /** @minLength 1 */
  bedNumber?: string;
  /** @minLength 1 */
  name?: string;
}

export interface MachineDeleteResult {
  deletedId: number;
}

export interface ProcessUpdate {
  done: boolean;
  /** @minLength 1 */
  updatedBy: string;
}

export interface BulkProcessInput {
  /** @minItems 1 */
  machineIds: number[];
  process: Process;
  done: boolean;
  /** @minLength 1 */
  updatedBy: string;
}

export interface BulkCompleteInput {
  /** @minItems 1 */
  machineIds: number[];
  /** @minLength 1 */
  updatedBy: string;
}

export type MachineSummaryProcessTotals = {
  reliability: number;
  laser: number;
  lkc: number;
  fuc: number;
  ct: number;
  tag: number;
};

export interface MachineSummary {
  total: number;
  notStarted: number;
  inProgress: number;
  completed: number;
  processTotals: MachineSummaryProcessTotals;
}

export interface Activity {
  id: number;
  machineId: number;
  bedNumber: string;
  machineName: string;
  process: Process;
  completed: boolean;
  updatedBy: string;
  updatedAt: string;
}

export interface ActivityClearResult {
  clearedCount: number;
}

export interface ErrorResponse {
  error: string;
}